Atualização

+++ Dtos/Pedidos/PedidoResposta.cs
@@ -11,10 +11,10 @@ namespace Biblioteca.Dtos.Pedidos
   {
     public int Id { get; set; }
 
-    public decimal Quantidade { get; set; }
+    public int Quantidade { get; set; }
     public int LivroId { get; set; }
     public LivroResposta Livro { get; set; }
-    public int UsuarioId{ get; set; }
-    public UsuarioResposta Usuario{ get; set; }
+    public int UsuarioId { get; set; }
+    public UsuarioResposta Usuario { get; set; }
   }
 }
